Move congestion_ctrl module (and directory) out of sender.
