+++ src/protocols/tcp/active_open.rs
@@ -10,7 +10,7 @@ use crate::{
         ip::IpProtocol,
         ipv4::{Ipv4Endpoint, Ipv4Header},
         tcp::{
-            established::cc::{self, CongestionControl},
+            established::congestion_ctrl::{self, CongestionControl},
             segment::{TcpHeader, TcpOptions2, TcpSegment},
         },
     },
@@ -215,7 +215,7 @@ impl<RT: Runtime> ActiveOpenSocket<RT> {
             tx_window_size,
             remote_window_scale,
             mss,
-            cc::None::new,
+            congestion_ctrl::None::new,
             None,
         );
         self.set_result(Ok(cb));
