+++ src/protocols/tcp/established/sender/mod.rs
@@ -1,11 +1,10 @@
 // Copyright (c) Microsoft Corporation.
 // Licensed under the MIT license.
 
-pub mod congestion_ctrl;
 mod rto;
 
-use self::{congestion_ctrl as cc, rto::RtoCalculator};
-use super::ControlBlock;
+use self::rto::RtoCalculator;
+use super::{congestion_ctrl, ControlBlock};
 use crate::protocols::tcp::{segment::TcpHeader, SeqNumber};
 use ::libc::{EBUSY, EINVAL};
 use ::runtime::{
@@ -86,7 +85,7 @@ pub struct Sender<RT: Runtime> {
     // Retransmission Timeout (RTO) calculator.
     pub rto: RefCell<RtoCalculator>,
 
-    pub congestion_ctrl: Box<dyn cc::CongestionControl<RT>>,
+    pub congestion_ctrl: Box<dyn congestion_ctrl::CongestionControl<RT>>,
 }
 
 impl<RT: Runtime> fmt::Debug for Sender<RT> {
@@ -110,8 +109,8 @@ impl<RT: Runtime> Sender<RT> {
         send_window: u32,
         window_scale: u8,
         mss: usize,
-        cc_constructor: cc::CongestionControlConstructor<RT>,
-        congestion_control_options: Option<cc::Options>,
+        cc_constructor: congestion_ctrl::CongestionControlConstructor<RT>,
+        congestion_control_options: Option<congestion_ctrl::Options>,
     ) -> Self {
         Self {
             send_unacked: WatchedValue::new(seq_no),
